+++ com/planet_ink/coffee_mud/Abilities/Prayers/Prayer_HolyDay.java
@@ -129,7 +129,7 @@ public class Prayer_HolyDay extends Prayer
 						godName="the gods";
 					R.showHappens(CMMsg.MSG_OK_VISUAL,"A holy day of "+godName+" has begun!");
 				}
-				beneficialAffect(mob,target,asLevel,(int)CommonStrings.getIntVar(CommonStrings.SYSTEMI_TICKSPERMUDDAY));
+				beneficialAffect(mob,target,asLevel,CommonStrings.getIntVar(CommonStrings.SYSTEMI_TICKSPERMUDDAY));
 			}
 		}
 		else
