+++ com/planet_ink/coffee_mud/Abilities/Songs/Dance_Clog.java
@@ -34,7 +34,7 @@ public class Dance_Clog extends Dance
 		if(invoker==null) return;
 		if(affected==invoker) return;
 
-		affectableStats.setAttackAdjustment((affectableStats.attackAdjustment()-invoker().charStats().getStat(CharStats.CHARISMA))-((int)Math.round(invoker.envStats().level())*2));
+		affectableStats.setAttackAdjustment((affectableStats.attackAdjustment()-invoker().charStats().getStat(CharStats.CHARISMA))-(Math.round(invoker.envStats().level())*2));
 	}
 	public void affectCharStats(MOB affected, CharStats affectableStats)
 	{
