+++ com/planet_ink/coffee_mud/Abilities/Songs/Dance_Polka.java
@@ -44,7 +44,7 @@ public class Dance_Polka extends Dance
 		if(invoker==null) return;
 		if(affected==invoker) return;
 
-		affectableStats.setStat(CharStats.DEXTERITY,(int)Math.round(affectableStats.getStat(CharStats.DEXTERITY)-3));
+		affectableStats.setStat(CharStats.DEXTERITY,Math.round(affectableStats.getStat(CharStats.DEXTERITY)-3));
 	}
 
 	public void show(MOB mob, int code, String text)
