+++ com/planet_ink/coffee_mud/Abilities/Songs/Song_Inebriation.java
@@ -43,7 +43,7 @@ public class Song_Inebriation extends Song
 		if(invoker==null) return;
 		if(affected==invoker) return;
 
-		affectableStats.setStat(CharStats.DEXTERITY,(int)Math.round(affectableStats.getStat(CharStats.DEXTERITY)-3));
+		affectableStats.setStat(CharStats.DEXTERITY,Math.round(affectableStats.getStat(CharStats.DEXTERITY)-3));
 	}
 
 	public void show(MOB mob, int code, String text)
