+++ com/planet_ink/coffee_mud/Abilities/Spells/Spell_Stoneskin.java
@@ -102,7 +102,7 @@ public class Spell_Stoneskin extends Spell
 			{
 				mob.location().send(mob,msg);
 				mob.location().show(target,null,CMMsg.MSG_OK_VISUAL,"<S-NAME> watch(es) <S-HIS-HER> skin turn hard as stone!");
-				HitsRemaining=5+(int)Math.round(adjustedLevel(mob,asLevel)/2);
+				HitsRemaining=5+Math.round(adjustedLevel(mob,asLevel)/2);
 				beneficialAffect(mob,target,asLevel,0);
 			}
 		}
