+++ com/planet_ink/coffee_mud/Items/MiscMagic/GenSuperPill.java
@@ -117,7 +117,7 @@ public class GenSuperPill extends GenPill
 
 		String val=Util.getParmStr(readableText,"gen","").toUpperCase();
 		if((val.length()>0)&&((val.charAt(0)=='M')||(val.charAt(0)=='F')||(val.charAt(0)=='N')))
-			mob.baseCharStats().setStat(CharStats.GENDER,(int)val.charAt(0));
+			mob.baseCharStats().setStat(CharStats.GENDER,val.charAt(0));
 		val=Util.getParmStr(readableText,"cla","").toUpperCase();
 		if((val.length()>0)&&(CMClass.getCharClass(val)!=null)&&(!val.equalsIgnoreCase("Archon")))
 		{
